# wip: EA overhaul