… Compiler/src/ssair/ir.jl
… Compiler/src/ssair/passes.jl
… Compiler/test/EAUtils.jl
… Compiler/test/irpasses.jl
… Compiler/test/irutils.jl
… doc/src/devdocs/EscapeAnalysis.md