This PR aims to implement the new design for EscapeAnalysis.jl as proposed in https://hackmd.io/XKTmg0R0Tt2giLW56mWZ3A. While the actual implementation deviates slightly from the design document, the high-level goals and concepts remain the same.

The current EscapeAnalysis.jl is based on the old escape analysis design of Java's Graal compiler[^2] and suffers from the following issues:
- Flow-insensitive analysis: To simplify the implementation, EscapeAnalysis.jl currently maintains and updates a single escape state for the entire method being analyzed, preventing flow-sensitive analysis.
- Backward analysis: To improve the efficiency of escape information propagation, the analysis is performed backward, but it suffers from handling alias information (i.e. information about object field), which inherently propagates forward.
- Sloppy convergence check: The current algorithm iterates over the IR until all states converge. This leads to unnecessary iterations even for simple cases.
- No support for inter-procedural alias analysis: Alias analysis is currently limited to intra-procedural contexts, significantly reducing its precision when encountering non-inlined calls.

The new EscapeAnalysis.jl takes inspiration from their (relatively) recent paper[^1] on partial escape analysis and adopts the following design:
- Flow-sensitive analysis: The new analysis, similar to our type inference, uses flow-sensitive abstract interpretation, allowing the propagation of escape/alias information based on control flow.
- Forward analysis: The new analysis focuses on propagating information along the control flow and is implemented as a forward analysis.
- Convergence check with a basic-block-based working set: Like our type inference, the new algorithm manages a working set of basic blocks during analysis iterations, enabling more efficient convergence checks.
- Inter-procedural alias analysis: By appropriately caching and propagating alias information inter-procedurally, the new design aims to preserve alias analysis precision even for IR containing non-inlined calls.

With this new analysis, a more powerful SROA (or load-forwarding) and a better type inference for capturing closure become possible. It may also enable optimizations like allocation sinking and stack allocation in the future.

As a showcase to the capabilities achieved by the new EscapeAnalysis.jl, the analysis results for the examples presented in their paper[^1] are as follows:
```julia
mutable struct Key
    idx::Int
    ref
    Key(idx::Int, @nospecialize(ref)) = new(idx, ref)
end
import Base: ==
key1::Key == key2::Key =
    key1.idx == key2.idx && key1.ref === key2.ref

global cache_key::Key
global cache_value

function get_value(idx::Int, ref)
    global cache_key, cache_value
    key = Key(idx, ref)
    if key == cache_key
        return cache_value
    else
        cache_key = key
        cache_value = create_value(key)
        return cache_value
    end
end
```
```julia
julia> result = code_escapes(get_value, (Int,Any))
get_value(✓ idx::Int64, X ref::Any) in Main at REPL[11]:1
3 1 ── X  %1  = %new(Main.Key, _2, _3)::Key                            │╻  Key
4 │    X  %2  = Main.cache_key::Key                                    │  
  │    ✓  %3  =   builtin Base.getfield(%1, :idx)::Int64 (↦ _2)        │╻  ==
  │    X  %4  =   builtin Base.getfield(%2, :idx)::Int64 (↦ X)         ││┃  getproperty
  │    ◌  %5  =   builtin (%3 === %4)::Bool                            ││╻  ==
  └─── ◌        goto #3 if not %5                                      ││ 
  2 ── X  %7  =   builtin Base.getfield(%1, :ref)::Any (↦ _3)          ││╻  getproperty
  │    X  %8  =   builtin Base.getfield(%2, :ref)::Any (↦ X)           │││
  │    ✓  %9  =   builtin (%7 === %8)::Bool                            ││ 
  └─── ◌        goto #4                                                ││ 
  3 ── ◌        goto #4                                                ││ 
  4 ┄─ ✓  %12 = φ (#2 => %9, #3 => false)::Bool                        │  
  └─── ◌        goto #6 if not %12                                     │  
5 5 ── X  %14 = Main.cache_value::Any                                  │  
  └─── ◌        return %14                                             │  
  6 ── ◌        nothing::Nothing                                       │  
  7 ── ◌        nothing::Nothing                                       │  
7 8 ── X          builtin Base.setglobal!(Main, :cache_key, %1)::Key   │  
8 │    X  %19 = Main.create_value::Any                                 │  
  │    X  %20 =   dynamic (%19)(%1)::Any                               │  
  │    X  %21 =   builtin Core.get_binding_type(Main, :cache_value)::Type 
  │    ◌  %22 =   builtin (%20 isa %21)::Bool                          │  
  └─── ◌        goto #10 if not %22                                    │  
  9 ── ◌        goto #11                                               │  
  10 ─ X  %25 =   dynamic Base.convert(%21, %20)::Any                  │  
  11 ┄ X  %26 = φ (#9 => %20, #10 => %25)::Any                         │  
  │    X          builtin Base.setglobal!(Main, :cache_value, %26)::Any│  
9 │    X  %28 = Main.cache_value::Any                                  │  
  └─── ◌        return %28                                             │

julia> EscapeAnalysis.has_no_escape(result.eresult.bbescapes[5][SSAValue(1)])
true
```
Here, `builtin Base.getfield(%1, :idx)::Int64 (↦ _2)` and `builtin Base.getfield(%1, :ref)::Any (↦ _3)` indicate that load-forwarding is possible for these `getfield` operations. Additionally, `EscapeAnalysis.has_no_escape(result.eresult.bbescapes[5][SSAValue(1)])` shows that `%1 = %new(Main.Key, _2, _3)::Key` does not (yet) escape in basic block 5, which demonstrates that allocation sinking optimization is possible for this IR.

While the new analysis captures significantly more information, its performance has not degraded substantially:
```julia
julia> @benchmark code_escapes(get_value, (Int,Any))
```
```julia
# master
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):   83.167 μs …   9.479 ms  ┊ GC (min … max): 0.00% … 97.99%
 Time  (median):      90.292 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   101.103 μs ± 200.798 μs  ┊ GC (mean ± σ):  6.07% ±  3.19%

   ▂▅██▇▆▆▅▅▄▄▃▃▂▂▂▂▁▁  ▁                                       ▂
  ▇█████████████████████████▆█▇▆▇▆▆▇▆▇▆▅▅▅▆▅▅▄▅▅▅▄▅▄▃▅▃▃▅▄▄▅▄▂▄ █
  83.2 μs       Histogram: log(frequency) by time        162 μs <

 Memory estimate: 135.00 KiB, allocs estimate: 2601.

# this PR
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):   89.750 μs …   9.834 ms  ┊ GC (min … max): 0.00% … 98.22%
 Time  (median):      95.625 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   105.633 μs ± 224.639 μs  ┊ GC (mean ± σ):  7.05% ±  3.48%

         ▁▆█▅▂                                                   
  ▁▁▁▂▂▂▅█████▇▅▄▄▅▄▄▃▃▃▃▃▃▃▃▃▂▂▂▂▂▂▂▂▂▂▂▂▁▂▁▁▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁ ▂
  89.8 μs          Histogram: frequency by time          120 μs <

 Memory estimate: 169.62 KiB, allocs estimate: 3172.
```
In particular, performance has improved for simpler cases:
```julia
julia> @benchmark code_escapes() do
           x = Ref{String}()
           x[] = "foo"
           out1 = x[]
           x[] = "bar"
           out2 = x[]
           return x, out1, out2
       end
```
```julia
# master
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  52.083 μs …   8.430 ms  ┊ GC (min … max): 0.00% … 98.20%
 Time  (median):     57.292 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   62.336 μs ± 162.891 μs  ┊ GC (mean ± σ):  5.72% ±  2.19%

               ▃█▇▃                                             
  ▁▁▁▁▁▁▂▄▅██▆▇████▆▄▄▄▄▄▄▄▃▄▄▄▃▃▃▃▃▂▃▂▂▂▂▂▂▂▂▂▂▂▁▁▁▁▁▁▁▁▁▁▁▁▁ ▂
  52.1 μs         Histogram: frequency by time         71.5 μs <

 Memory estimate: 85.12 KiB, allocs estimate: 1890.

# this PR
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  44.750 μs …   9.561 ms  ┊ GC (min … max): 0.00% … 98.61%
 Time  (median):     50.938 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   62.723 μs ± 219.193 μs  ┊ GC (mean ± σ):  5.54% ±  1.97%

   ██▇▅▁                                                        
  ▅█████▇▅▄▃▃▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▁▂▁▂▂▂▂▁▂▂▁▂▂▂▁▁▂▂▂▂▁▂ ▃
  44.8 μs         Histogram: frequency by time          147 μs <

 Memory estimate: 77.66 KiB, allocs estimate: 1688.
```

Furthermore, while their paper does not discuss inter-procedural alias analysis in much detail, this PR also aims to implement inter-procedural alias analysis. Ultimately, the goal is to successfully analyze targets such as the following:
```julia
function issue56561_2(a)
    x = sin(a) x = identity(x)
    return (a -> a + x)(a)
end
```

This PR is still a work in progress, and the following TODO list outlines the remaining tasks:
1. [x] Switch to the new flow-sensitive and working-set-based algorithm  
2. [x] Implement the new alias analysis design (locally)  
3. [ ] Reimplement inter-procedural escape information propagation  
4. [ ] Implement the inter-procedural alias analysis design  
5. [ ] Optimize performance  
6. [ ] Enhance `sroa_mutables!` using EscapeAnalysis  
    1. [ ] Make `MemoryInfo` CFG-aware  
7. [ ] Enable optimizations for capturing closures in combination with JuliaLang/julia#56687  
(the actual application to optimizations (6–7) may be addressed in a separate PR)

[^1]: Lukas Stadler, Thomas Würthinger, and Hanspeter Mössenböck. 2018. Partial Escape Analysis and Scalar Replacement for Java. In Proceedings of Annual IEEE/ACM International Symposium on Code Generation and Optimization (CGO '14). Association for Computing Machinery, New York, NY, USA, 165–174. https://doi.org/10.1145/2544137.2544157

[^2]: Thomas Kotzmann and Hanspeter Mössenböck. 2005. Escape analysis in the context of dynamic compilation and deoptimization. In Proceedings of the 1st ACM/USENIX international conference on Virtual execution environments (VEE '05). Association for Computing Machinery, New York, NY, USA, 111–120. https://doi.org/10.1145/1064979.1064996