@aviatesk 2024-12-17
The `EscapeResult` data structure is closely tied to this overhaul, so it’s difficult to separate them.