@aviatesk 2024-12-19
Of course it is possible to perform the rename, but the name `EscapeResult` in the previous code did not hold much meaning. This is because the previous analysis managed and updated a single state, which could then be returned directly as the final result. In that sense, `EscapeState == EscapeResult`.
In the analysis being implemented now, multiple `BlockEscapeState` structures manage the analysis state for each basic block. These are distinct from the final analysis result, `EscapeResult`, which represents the escape information for the entire method. So the rename was necessary and kind of tied to this refactoring.

That said, if it makes the review process easier, it might be worth trying.