@aviatesk 2024-12-19
@nanosoldier `runbenchmarks("inference", vs=":master")`